+++ src/sns/report/model/sorts.rs
@@ -64,6 +64,35 @@ impl SnsProposalStatusFilter {
 }
 
 #[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
+pub enum SnsProposalTopicFilter {
+    #[default]
+    Any,
+    DaoCommunitySettings,
+    SnsFrameworkManagement,
+    DappCanisterManagement,
+    ApplicationBusinessLogic,
+    Governance,
+    TreasuryAssetManagement,
+    CriticalDappOperations,
+}
+
+impl SnsProposalTopicFilter {
+    #[must_use]
+    pub const fn as_str(self) -> &'static str {
+        match self {
+            Self::Any => "any",
+            Self::DaoCommunitySettings => "dao-community-settings",
+            Self::SnsFrameworkManagement => "sns-framework-management",
+            Self::DappCanisterManagement => "dapp-canister-management",
+            Self::ApplicationBusinessLogic => "application-business-logic",
+            Self::Governance => "governance",
+            Self::TreasuryAssetManagement => "treasury-asset-management",
+            Self::CriticalDappOperations => "critical-dapp-operations",
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
 pub enum SnsListSort {
     #[default]
     Id,